feat: add hooks to the `Immer` class
Hooks are passed as options to the Immer class.

onAssign(state, prop, value)
  Called for every property assigned to by a producer

onDelete(state, prop)
  Called for every property deleted by a producer

onCopy(state)
  Called whenever a copy becomes finalized

These hooks are passed an `ImmerState` object. (defined at the end of "src/index.d.ts")

The `state` is guaranteed to be both modified and finalized if ever passed to a hook.