fix: produce no longer accepts non draftable objects, fixes #596
BREAKING CHANGE: `produce` no longer accepts any base state to operate on that is not draftable, something which caused quite some confusion in the past. If you were using `produce` in a way where it might be invoked on non draftable data, you need to guard it with `isDraftable` from now before invoking the producer.