non-enumerable (immutable) methods on Map/Set
when an immer object includes a Map/Set, the mutating methods are replaced with ones that prevent mutation outside of the procude function.  
to prevent cluttering the console when inspecting the Map/Set, these methods are set to non-enumerable. this behavior also matches that of the original methods that are replaced.