Update curried-produce.mdx
First, thanks for this library.

I hope you'll find this slight rewriting accepatble.

On first reading, I found https://immerjs.github.io/immer/curried-produce extremely confusing.
The main reason being that, when I read "Passing a function as the first argument", I thought you meant that we were seeing a variation of the `produce` from the previous page, where both arguments were functions. I believe it's important to note that we'll see a very different way to call `produce`.
It's even more confusing, because you introduce a new way to call `produce` with an argument, and your example does not use this new way. I spend quite some time trying to understand how this example works. I believe it is better to explain to the user that they are going to see the "bad" version of the example first. Then explain what's wrong with this code (too much boilerplate), and how it can be improved. Finally, once they saw the nice code, explain how it work.


I also suspect that, in order to emphasize that `toggleTodo` consists simply of `produce` applied to the function that implements the actual business logic, it's nice to add this business logic as a separate function. This way, we get the very simple `const toggleTodo = produce(toggleTodo_)`. I don't state this would be good production code; I admit that in real life, an anonymous lambda would be better. However, for educational purpose, I find this clearer. This really emphasize that `toggleTodo` is semantically the exact same function as `toggleTodo_` except that it does not uses side effect but output a new value.